bluetooth: btfm_slim: Fix Clang 21 -Wdefault-const-init-var-unsafe warning
This commit addresses a compilation warning in `drivers/bluetooth/btfm_slim.c`
when building with Clang 21:
`default initialization of an object of type 'typeof (timeout)' (aka 'const unsigned long') leaves the object uninitialized [-Werror,-Wdefault-const-init-var-unsafe]`

The warning occurs on line 307 within a `time_before` macro call.
The `timeout` variable is a `const unsigned long`, and Clang's `typeof`
operator, used internally by `time_before`, flags an implicit uninitialized
variable creation.

To resolve this, the `timeout` variable is now explicitly cast to `unsigned long`
before being passed to `time_before`. This ensures that the type checking
within the macro is satisfied and suppresses the warning, while maintaining
the correct program logic.

+++ drivers/bluetooth/btfm_slim.c
@@ -304,7 +304,7 @@ static int btfm_slim_get_logical_addr(struct slim_device *slim)
 		/* Give SLIMBUS time to report present and be ready. */
 		usleep_range(1000, 1100);
 		BTFMSLIM_DBG("retyring get logical addr");
-	} while (time_before(jiffies, timeout));
+	} while (time_before(jiffies, (unsigned long)timeout));
 
 	return ret;
 }
